#016.Confirmation pop-up on Delete click added

+++ web-customer-tracker/WebContent/WEB-INF/view/list-customers.jsp
@@ -60,7 +60,7 @@
 						<td> 
 							<a href="${updateLink}">Update</a>
 							|
-							<a href="${deleteLink}">Delete</a>
+							<a href="${deleteLink}" onclick="if(!(confirm('Are you sure you want to delete this customer?'))) return false">Delete</a>
 						 </td>
 					</tr>
